[FEAT] PostService 게시글 삭제 메소드 수정

+++ src/main/java/hello/golong/domain/post/application/PostService.java
@@ -112,7 +112,10 @@ public class PostService {
         if(postOptional.isPresent()) {
             postRepository.deleteById(post_id);
             imgService.deleteImg(post_id, 0L);
-            reviewService.deleteReview(reviewService.findReviewByPostId(post_id).getId());
+            if(postOptional.get().getStatus() == 4) {
+                reviewService.deleteReview(reviewService.findReviewByPostId(post_id).getId());
+            }
+
         }
     }
